Extracted Customer and transaction and fixed linking issues
