+++ ekspedientki/.vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "files.associations": {
-        "random": "c"
+        "random": "c",
+        "clerk.h": "c"
     }
 }
